+++ react-rs/.gitignore
@@ -1,4 +1,3 @@
-# Logs
 logs
 *.log
 npm-debug.log*
@@ -12,7 +11,6 @@ dist
 dist-ssr
 *.local
 
-# Editor directories and files
 .vscode/*
 !.vscode/extensions.json
 .idea
@@ -22,3 +20,6 @@ dist-ssr
 *.njsproj
 *.sln
 *.sw?
+
+.next
+next-env.d.ts
