+++ react-rs/next.config.mjs
@@ -1,0 +1,7 @@
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+    output: 'export', 
+    distDir: './dist'
+  }
+   
+  export default nextConfig
