+++ react-rs/package.json
@@ -4,8 +4,9 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "dev": "vite",
-    "build": "tsc -b && vite build",
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives",
     "format:fix": "prettier --write \"**/*.{js,jsx,ts,tsx,json,css}\"",
     "preview": "vite preview",
@@ -15,10 +16,10 @@
   "dependencies": {
     "@reduxjs/toolkit": "^2.2.6",
     "@types/react-redux": "^7.1.33",
+    "next": "^14.2.5",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-redux": "^9.1.2",
-    "react-router-dom": "^6.24.1"
+    "react-redux": "^9.1.2"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^6.4.8",
@@ -43,7 +44,6 @@
     "prettier": "3.3.2",
     "redux-mock-store": "^1.5.4",
     "ts-jest": "^29.2.2",
-    "typescript": "^5.5.3",
-    "vite": "^5.3.1"
+    "typescript": "^5.5.3"
   }
 }
