+++ react-rs/src/ThemeContext.tsx
@@ -1,4 +1,33 @@
-import React, { createContext, useState, ReactNode } from "react";
+// import React, { createContext, useState, ReactNode } from "react";
+
+// interface ThemeContextType {
+//   theme: string;
+//   toggleTheme: () => void;
+// }
+
+// export const ThemeContext = createContext<ThemeContextType | undefined>(
+//   undefined,
+// );
+
+// export const ThemeProvider: React.FC<{ children: ReactNode }> = ({
+//   children,
+// }) => {
+//   const [theme, setTheme] = useState("light");
+
+//   const toggleTheme = () => {
+//     setTheme((prevTheme) => (prevTheme === "light" ? "dark" : "light"));
+//   };
+
+//   return (
+//     <ThemeContext.Provider value={{ theme, toggleTheme }}>
+//       <div className="wrapper" data-theme={theme}>
+//         {children}
+//       </div>
+//     </ThemeContext.Provider>
+//   );
+// };
+
+import React, { createContext, useState, ReactNode, useEffect } from "react";
 
 interface ThemeContextType {
   theme: string;
@@ -14,15 +43,17 @@ export const ThemeProvider: React.FC<{ children: ReactNode }> = ({
 }) => {
   const [theme, setTheme] = useState("light");
 
+  useEffect(() => {
+    document.body.dataset.theme = theme;
+  }, [theme]);
+
   const toggleTheme = () => {
     setTheme((prevTheme) => (prevTheme === "light" ? "dark" : "light"));
   };
 
   return (
     <ThemeContext.Provider value={{ theme, toggleTheme }}>
-      <div className="wrapper" data-theme={theme}>
-        {children}
-      </div>
+      <div data-theme={theme}>{children}</div>
     </ThemeContext.Provider>
   );
 };
